fix: make repositories route a layout to display child routes
Split `repositories.tsx` into a layout route (renders `Outlet`) and
`repositories.index.tsx` (renders the repo list). Previously, the
`/repositories` route rendered `RepositoriesPage` directly without an
`<Outlet />`, so navigating to `/repositories/$repoId` changed the URL
but never displayed the branch list.
